added cascade,removed the global hibernateSequence

+++ src/main/java/com/shikhakunj/udemymicroservicestutorial/user/jpa/UserJpaController.java
@@ -74,6 +74,10 @@ public class UserJpaController {
 	
 	@DeleteMapping("/jpa/users/{id}")
 	public void deleteUserById(@PathVariable int id) {
+		Optional<User> user = userRepo.findById(id);
+		if(user.isEmpty()){
+			throw new UserNotFoundException("id: "+id);
+		}
 		userRepo.deleteById(id);
 	}
